+++ packages/grid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.6.0 (2026-08-26)
+
 ### Internal
 
 -   Rename CSS module class names to kebab-case to satisfy `selector-class-pattern`. ([#82053](https://github.com/WordPress/gutenberg/pull/82053))